% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpal2-package.R
\docType{package}
\name{ggpal2-package}
\alias{ggpal2-package}
\title{ggpal2: "chores" extension ggplot2 LLM assistant}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extension of "chores" FKA "pals", a LLM assistant add-in for RStudio and Positron IDEs built on "ellmer". "ggpal2" is an assistant designed to help with ggplot2.
}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

}
\keyword{internal}
