% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-replace.R
\name{ggplot2-replace}
\alias{ggplot2-replace}
\alias{ggpal2}
\alias{ggplot2}
\title{the ggplot2 pal (ggplot2-replace)}
\source{
{pal} https://simonpcouch.github.io/pal/articles/pal.html
}
\description{
how-to use "ggplot2", the ggplot2-replace pal,
}
\section{instructions}{


Pals are interfaced with the via the pal addin.
For easiest access,
we recommend registering the pal addin to a keyboard shortcut.

i.e. for RStudio IDE, \verb{Tools > Modify Keyboard Shortcuts > Search "Pal"}
we suggest \verb{Ctrl+Alt+P (or Ctrl+Cmd+P on macOS).}

for more information, follow the {pal} instructions at https://simonpcouch.github.io/pal/articles/pal.html
}

\keyword{ggplot2-replace}
